# public/doc_swagger/openapi.yaml
openapi: 3.0.0
info:
  title: 'Api oAuth 2.0'
  description: 'Esta é a documentação base da sua API. Ela contém todas as informações necessárias para que você, sua equipe ou um desenvolvedor possam integrar o seu sistema com outros sistemas, sites, portais ou aplicativos móveis (iOS/Android)'
  version: 2.0.1
paths:
...
  /api/produtos:
    get:
      tags:
        - Produtos
      summary: Listar
      operationId: 4e44f31c7df5e5da4bb7c853e5396ae1
      parameters:
        -
          name: Authorization
          in: header
          required: true
          schema:
            description: 'Bearer token para autorização'
            type: string
            default: 'Bearer <token>'
      responses:
        '200':
          description: ''
    post:
      tags:
        - Produtos
      summary: Cadastrar
      operationId: c0714a29fe17a046dd3aa260961d54c1
      parameters:
        -
          name: Authorization
          in: header
          required: true
          schema:
            description: 'Bearer token para autorização'
            type: string
            default: 'Bearer <token>'
      requestBody:
        description: 'Input data format'
        content:
          multipart/form-data:
            schema:
              properties:
                produto:
                  description: Produto
                  type: string
              type: object
      responses:
        '201':
          description: ''
  '/api/produtos/{id}':
    get:
      tags:
        - Produtos
      summary: Retornar
      operationId: 75f25eff9644d3e89f7d18e3a8036e35
      parameters:
        -
          name: Authorization
          in: header
          required: true
          schema:
            description: 'Bearer token para autorização'
            type: string
            default: 'Bearer <token>'
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
    put:
      tags:
        - Produtos
      summary: Atualizar
      operationId: a319c3baf11db9a513306549d761a8af
      parameters:
        -
          name: Authorization
          in: header
          required: true
          schema:
            description: 'Bearer token para autorização'
            type: string
            default: 'Bearer <token>'
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: 'Input data format'
        content:
          multipart/form-data:
            schema:
              properties:
                produto:
                  description: Produto
                  type: string
              type: object
      responses:
        '201':
          description: ''
    delete:
      tags:
        - Produtos
      summary: Deletar
      operationId: 53c611ff7b82b5218a7fc59621aaab00
      parameters:
        -
          name: Authorization
          in: header
          required: true
          schema:
            description: 'Bearer token para autorização'
            type: string
            default: 'Bearer <token>'
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
...